manual - correcting a link, adding missing step
Manual installation has a wrong link to the zip archive (for some reason it displays https://github.com/revolunet/sublimetext-markdown-preview/archive/master.zip), correcting it now. There is also a missing 4th step it seems to me - one needs to extract the zip archive and move it to Packages folder, for me ST3 does not do that automatically.